+++ apps/settings/serializers/__init__.py
@@ -4,11 +4,11 @@
 from .auth import *
 from .basic import *
 from .cleaning import *
-from .email import *
+from .feature import *
+from .msg import *
+from .msg import *
 from .other import *
 from .public import *
 from .security import *
 from .settings import *
 from .terminal import *
-from .vault import *
-
